removed tests.py and added test_user.py
